[Moments] Many bug fixes and checks
Before we had a discrepancy that std.dev. from moments was 50% too
large, now it is only 2/3 of the result from quadrature of evaluating
P(Q) at various Q values. The discrepancy is not due to problems in
integrating the moments though there were many errors before. I
checked that without integration, I reproduce compound Poisson
results. Perhaps the posterior average of the second moment is
typically too small so the std.dev is underestimated